GO-3795 Fix comments
