+++ util/pbtypes/pbtypes_test.go
@@ -177,68 +177,3 @@ func TestIsEmptyValueOrAbsent(t *testing.T) {
 		})
 	}
 }
-
-func TestAddRemoveValue(t *testing.T) {
-	newStruct := func() *types.Struct {
-		return &types.Struct{Fields: map[string]*types.Value{
-			"tag":     StringList([]string{"red", "black"}),
-			"author":  String("William Shakespeare"),
-			"haters":  StringList([]string{}),
-			"year":    Int64(1564),
-			"numbers": IntList(8, 13, 21, 34),
-		}}
-	}
-
-	for _, tc := range []struct {
-		name     string
-		key      string
-		s        *types.Struct
-		toAdd    *types.Value
-		expected *types.Value
-	}{
-		{"string list + string list", "tag", newStruct(), StringList([]string{"blue", "green"}), StringList([]string{"red", "black", "blue", "green"})},
-		{"string list + string list (intersect)", "tag", newStruct(), StringList([]string{"blue", "black"}), StringList([]string{"red", "black", "blue"})},
-		{"string + string list", "author", newStruct(), StringList([]string{"Victor Hugo"}), StringList([]string{"William Shakespeare", "Victor Hugo"})},
-		{"string list + string", "tag", newStruct(), String("orange"), StringList([]string{"red", "black", "orange"})},
-		{"int list + int list", "numbers", newStruct(), IntList(55, 89), IntList(8, 13, 21, 34, 55, 89)},
-		{"int list + int list (intersect)", "numbers", newStruct(), IntList(13, 8, 55), IntList(8, 13, 21, 34, 55)},
-		{"int + int list", "year", newStruct(), IntList(1666, 2025), IntList(1564, 1666, 2025)},
-		{"int list + int", "numbers", newStruct(), Int64(55), IntList(8, 13, 21, 34, 55)},
-		{"string list + empty", "haters", newStruct(), StringList([]string{"Tomas River", "Leo Tolstoy"}), StringList([]string{"Tomas River", "Leo Tolstoy"})},
-		{"string list + no such key", "plays", newStruct(), StringList([]string{"Falstaff", "Romeo and Juliet", "Macbeth"}), StringList([]string{"Falstaff", "Romeo and Juliet", "Macbeth"})},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			AddValue(tc.s, tc.key, tc.toAdd)
-			assert.True(t, Get(tc.s, tc.key).Equal(tc.expected))
-		})
-	}
-
-	for _, tc := range []struct {
-		name     string
-		key      string
-		s        *types.Struct
-		toRemove *types.Value
-		expected *types.Value
-	}{
-		{"string list - string list", "tag", newStruct(), StringList([]string{"red", "black"}), nil},
-		{"string list - string list (some are not presented)", "tag", newStruct(), StringList([]string{"blue", "black"}), StringList([]string{"red"})},
-		{"string list - string", "tag", newStruct(), String("red"), StringList([]string{"black"})},
-		{"string - string list", "author", newStruct(), StringList([]string{"William Shakespeare"}), StringList([]string{})},
-		{"int list - int list", "numbers", newStruct(), IntList(13, 34), IntList(8, 21)},
-		{"int list - int list (some are not presented)", "numbers", newStruct(), IntList(2020, 5), IntList(8, 13, 21, 34)},
-		{"int - int list", "year", newStruct(), IntList(1380, 1564), IntList()},
-		{"int list - int", "numbers", newStruct(), Int64(21), IntList(8, 13, 34)},
-		{"empty - string list", "haters", newStruct(), StringList([]string{"Tomas River", "Leo Tolstoy"}), StringList([]string{})},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			RemoveValue(tc.s, tc.key, tc.toRemove)
-			assert.True(t, Get(tc.s, tc.key).Equal(tc.expected))
-		})
-	}
-}
-
-func TestEqual(t *testing.T) {
-	v1 := String("red")
-	v2 := String("red")
-	assert.True(t, v1.Equal(v2))
-}
